+++ api/org.integratedmodelling.klab.api/src/org/integratedmodelling/klab/api/monitoring/IMessage.java
@@ -204,7 +204,8 @@ public interface IMessage {
 		/*
 		 * F->B: ask engine to modify or delete projects or project assets
 		 */
-		CreateNamespace, CreateScenario, DeleteNamespace, DeleteLocalResource, CreateProject, DeleteProject, CreateScript, DeleteScript, CreateTestCase, DeleteTestCase,
+		CreateNamespace, CreateScenario, DeleteNamespace, DeleteLocalResource, CreateProject, DeleteProject,
+		CreateScript, DeleteScript, CreateTestCase, DeleteTestCase,
 
 		/**
 		 * F->B: notification when files are explicitly changed, added or deleted;
@@ -249,6 +250,12 @@ public interface IMessage {
 		ExecuteObservationAction,
 
 		/**
+		 * F->B Authorization class - inquiries about permitted operations and network
+		 * status
+		 */
+		NetworkStatus,
+
+		/**
 		 * --- Task lifecycle --- B -> F
 		 */
 		ScriptStarted, TaskStarted, TaskFinished, TaskAborted, DataflowCompiled, DataflowStateChanged,
@@ -257,11 +264,11 @@ public interface IMessage {
 		 * Task lifecycle F -> B
 		 */
 		TaskInterrupted, DataflowNodeDetail, DataflowNodeRating,
-		
+
 		/**
 		 * Scheduler lifecycle F->B
 		 */
-		SchedulingStarted, SchedulingFinished, SchedulerReset, 
+		SchedulingStarted, SchedulingFinished, SchedulerReset,
 
 		/*
 		 * --- Search-class types ---
@@ -286,7 +293,8 @@ public interface IMessage {
 		/*
 		 * --- ResourceLifecycle-class types, F->B
 		 */
-		ImportResource, DeleteResource, UpdateResource, ValidateResource, PreviewResource, CopyResource, MoveResource, CreateResource, ImportIntoResource, ResourceOperation,
+		ImportResource, DeleteResource, UpdateResource, ValidateResource, PreviewResource, CopyResource, MoveResource,
+		CreateResource, ImportIntoResource, ResourceOperation,
 
 		/*
 		 * --- ResourceLifecycle-class types, B->F