+++ klab.engine/src/main/java/org/integratedmodelling/klab/auth/KlabCertificate.java
@@ -164,7 +164,7 @@ public class KlabCertificate implements ICertificate {
 			return true;
 		}
 
-		return false;
+		return true;
 	}
 
 	/**