+++ klab.engine/src/main/java/org/integratedmodelling/klab/auth/Node.java
@@ -23,6 +23,8 @@ public class Node implements INodeIdentity {
     private String authenticatingHub;
     private List<String> urls = new ArrayList<>();
     private Set<Permission> permissions = new HashSet<>();
+    private Set<String> resourceUrls = new HashSet<>();
+    private Set<String> adapterIds = new HashSet<>();
     private Date bootTime = new Date();
     private String token;
     private boolean online;
@@ -42,6 +44,7 @@ public class Node implements INodeIdentity {
 		this.urls.addAll(node.getUrls());
 		this.parent = partner;
 		this.token = token;
+		this.adapterIds.addAll(node.getAdapters());
 		// TODO permissions
     }
 
@@ -177,6 +180,16 @@ public class Node implements INodeIdentity {
 
 	public void setBootTime(Date bootTime) {
 		this.bootTime = bootTime;
+	}
+
+	@Override
+	public Set<String> getAdapters() {
+		return adapterIds;
+	}
+
+	@Override
+	public Set<String> getResources() {
+		return resourceUrls;
 	}