+++ klab.engine/src/main/java/org/integratedmodelling/klab/engine/rest/controllers/network/NetworkController.java
@@ -11,7 +11,9 @@ import org.springframework.web.bind.annotation.RestController;
  *
  */
 @RestController
-@Secured(Roles.OWNER)
+@Secured({ Roles.ENGINE_USER, Roles.ADMIN })
 public class NetworkController {
- 
+
+	
+	
 }