+++ klab.engine/src/main/java/org/integratedmodelling/klab/engine/runtime/Session.java
@@ -32,12 +32,15 @@ import org.integratedmodelling.klab.Indexing;
 import org.integratedmodelling.klab.Klab;
 import org.integratedmodelling.klab.Logging;
 import org.integratedmodelling.klab.Namespaces;
+import org.integratedmodelling.klab.Network;
 import org.integratedmodelling.klab.Observables;
 import org.integratedmodelling.klab.Observations;
 import org.integratedmodelling.klab.Resources;
 import org.integratedmodelling.klab.Units;
 import org.integratedmodelling.klab.api.auth.IEngineUserIdentity;
 import org.integratedmodelling.klab.api.auth.IIdentity;
+import org.integratedmodelling.klab.api.auth.INetworkSessionIdentity;
+import org.integratedmodelling.klab.api.auth.INodeIdentity;
 import org.integratedmodelling.klab.api.auth.IRuntimeIdentity;
 import org.integratedmodelling.klab.api.auth.Roles;
 import org.integratedmodelling.klab.api.data.CRUDOperation;
@@ -87,6 +90,9 @@ import org.integratedmodelling.klab.rest.DataflowDetail;
 import org.integratedmodelling.klab.rest.DataflowState;
 import org.integratedmodelling.klab.rest.DocumentationReference;
 import org.integratedmodelling.klab.rest.InterruptTask;
+import org.integratedmodelling.klab.rest.NetworkReference;
+import org.integratedmodelling.klab.rest.NodeReference;
+import org.integratedmodelling.klab.rest.NodeReference.Permission;
 import org.integratedmodelling.klab.rest.ObservableReference;
 import org.integratedmodelling.klab.rest.ObservationRequest;
 import org.integratedmodelling.klab.rest.ProjectLoadRequest;
@@ -470,6 +476,33 @@ public class Session implements ISession, UserDetails, IMessageBus.Relay {
 		}
 	}
 
+	@MessageHandler(messageClass = IMessage.MessageClass.Authorization, type = IMessage.Type.NetworkStatus)
+	private void handleNetworkStatusRequest(String dummy) {
+		/*
+		 * send back a network descriptor with all nodes we can publish to at the moment
+		 * of the call.
+		 */
+		NetworkReference ret = new NetworkReference();
+		ret.setHub(Network.INSTANCE.getHub());
+		INetworkSessionIdentity network = this.getParentIdentity(INetworkSessionIdentity.class);
+		if (network != null) {
+			for (INodeIdentity node : network.getNodes()) {
+				NodeReference desc = Network.INSTANCE.getNodeDescriptor(node.getName());
+				if (desc != null) {
+					if (node.getPermissions().contains(Permission.PUBLISH)) {
+						ret.getPublishing().add(node.getName());
+					}
+					if (node.getPermissions().contains(Permission.QUERY)) {
+						ret.getSearchable().add(node.getName());
+					}
+					desc.getAdapters().addAll(node.getAdapters());
+					ret.getNodes().put(node.getName(), desc);
+				}
+			}
+		}
+		monitor.send(IMessage.MessageClass.Authorization, IMessage.Type.NetworkStatus, ret);
+	}
+
 	@MessageHandler(type = IMessage.Type.FeatureAdded)
 	private void handleFeatureAdded(final SpatialLocation location) {
 
@@ -586,7 +619,7 @@ public class Session implements ISession, UserDetails, IMessageBus.Relay {
 					}
 				}.start();
 			}
-			
+
 		} else if (type == IMessage.Type.ImportIntoResource) {
 
 			IResource resource = Resources.INSTANCE.resolveResource(request.getTargetResourceUrn());
@@ -773,8 +806,8 @@ public class Session implements ISession, UserDetails, IMessageBus.Relay {
 						for (ObservableReference observable : Authentication.INSTANCE
 								.getDefaultObservables(Session.this)) {
 							SearchMatch match = new SearchMatch(observable.getObservable(), observable.getLabel(),
-									observable.getDescription(), observable.getSemantics(),
-									observable.getState(), observable.getExtendedDescription());
+									observable.getDescription(), observable.getSemantics(), observable.getState(),
+									observable.getExtendedDescription());
 							match.setIndex(i++);
 							response.getMatches().add(match);
 							matches.add(new org.integratedmodelling.klab.engine.indexing.SearchMatch(match));